docs: start to setup versioned docs
