+++ documentation/docs/4/06-legacy/01-svelte-register.md
