+++ documentation/docs/5/01-getting-started/01-introduction.md
@@ -1,0 +1,7 @@
+---
+title: Introduction
+---
+
+You can find the latest Svelte reference documentation at [https://svelte.dev/docs](https://svelte.dev/docs).
+
+You can learn more about Svelte 5 and other future work on our [roadmap](https://svelte.dev/roadmap).
