+++ documentation/docs/5/01-getting-started/meta.json
@@ -1,0 +1,3 @@
+{
+	"title": "Getting Started"
+}
