+++ sites/svelte.dev/src/routes/docs/5/+page.svelte
@@ -1,0 +1,7 @@
+<script>
+	import Docs from '../[slug]/Docs.svelte';
+
+	export let data;
+</script>
+
+<Docs version={5} {data} />
